fix: do not delete the tns tsconfig for shared apps as it contains a lot of custom code.

+++ lib/controllers/migrate-controller.ts
@@ -129,7 +129,10 @@ export class MigrateController extends UpdateControllerBase implements IMigrateC
 		this.$fs.deleteDirectory(path.join(projectData.projectDir, constants.NODE_MODULES_FOLDER_NAME));
 		this.$fs.deleteFile(path.join(projectData.projectDir, constants.WEBPACK_CONFIG_NAME));
 		this.$fs.deleteFile(path.join(projectData.projectDir, constants.PACKAGE_LOCK_JSON_FILE_NAME));
-		this.$fs.deleteFile(path.join(projectData.projectDir, constants.TSCCONFIG_TNS_JSON_NAME));
+		if (!projectData.isShared) {
+			this.$fs.deleteFile(path.join(projectData.projectDir, constants.TSCCONFIG_TNS_JSON_NAME));
+		}
+
 		this.$logger.info("Clean old project artefacts complete.");
 	}
